providers/openai/01.content_gen.py
import os
from dotenv import load_dotenv
from openai import OpenAI
from openai.types.chat.chat_completion import ChatCompletion

load_dotenv()

## https://ai.google.dev/gemini-api/docs/openai


client = OpenAI(
    api_key=os.environ.get("GEMINI_API_KEY"),
    base_url="https://generativelanguage.googleapis.com/v1beta/openai/",
)


response: ChatCompletion = client.chat.completions.create(
    model="gemini-3-flash-preview",
    messages=[
        {"role": "system", "content": "You are a helpful assistant."},
        {"role": "user", "content": "Please reply ok"},
    ],
    temperature=0.1,
)

print("ai:", response.choices[0].message.content)
